# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Bharat-Automator OS — Unified Agentic Mesh API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finance/file-gst:
    post:
      operationId: fileGST
      tags: [finance]
      summary: Autonomously file GST return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GSTFilingRequest"
      responses:
        "200":
          description: GST filing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxFilingResult"
  /finance/file-income-tax:
    post:
      operationId: fileIncomeTax
      tags: [finance]
      summary: Autonomously file Income Tax return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeTaxRequest"
      responses:
        "200":
          description: Income tax filing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxFilingResult"
  /finance/bid-freelance:
    post:
      operationId: bidFreelance
      tags: [finance]
      summary: Autonomously bid on freelance jobs (Contra/Truelancer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelanceBidRequest"
      responses:
        "200":
          description: Freelance bid result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelanceBidResult"
...
  /indiastack/auth-aadhaar:
    post:
      operationId: authenticateAadhaar
      tags: [indiastack]
      summary: Authenticate citizen via Aadhaar OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AadhaarAuthRequest"
      responses:
        "200":
          description: Authentication result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AadhaarAuthResult"
  /indiastack/upi-payment:
    post:
      operationId: initiateUpiPayment
      tags: [indiastack]
      summary: Initiate a UPI payment for agent-executed transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpiPaymentRequest"
      responses:
        "200":
          description: Payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiPaymentResult"
  /indiastack/digilocker-docs:
    get:
      operationId: fetchDigilockerDocs
      tags: [indiastack]
      summary: Fetch citizen documents from DigiLocker
      parameters:
        - name: citizenId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigiLockerDocs"
...
components:
  schemas:
...
    GSTFilingRequest:
      type: object
      properties:
        gstin:
          type: string
        period:
          type: string
        invoiceData:
          type: array
          items:
            type: object
            additionalProperties: true
        bankAccountLinked:
          type: boolean
      required: [gstin, period, invoiceData, bankAccountLinked]
    IncomeTaxRequest:
      type: object
      properties:
        panNumber:
          type: string
        assessmentYear:
          type: string
        incomeDetails:
          type: object
          additionalProperties: true
        form16Available:
          type: boolean
      required: [panNumber, assessmentYear, incomeDetails, form16Available]
    TaxFilingResult:
      type: object
      properties:
        referenceNumber:
          type: string
        status:
          type: string
          enum: [filed, pending_verification, error, resubmission_required]
        filedAt:
          type: string
        acknowledgementNumber:
          type: string
        refundAmount:
          type: number
        nextSteps:
          type: array
          items:
            type: string
      required: [referenceNumber, status, filedAt, nextSteps]
    FreelanceBidRequest:
      type: object
      properties:
        studentId:
          type: string
        skills:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
            enum: [Contra, Truelancer, Upwork, Fiverr]
        minBudget:
          type: number
        maxBudgetPerProject:
          type: number
        portfolioUrl:
          type: string
      required: [studentId, skills, platforms, minBudget]
    FreelanceBidResult:
      type: object
      properties:
        bidsPlaced:
          type: integer
        platforms:
          type: array
          items:
            type: string
        jobsTargeted:
          type: array
          items:
            $ref: "#/components/schemas/JobBid"
        estimatedResponseTime:
          type: string
      required: [bidsPlaced, platforms, jobsTargeted, estimatedResponseTime]
    JobBid:
      type: object
      properties:
        jobId:
          type: string
        title:
          type: string
        platform:
          type: string
        bidAmount:
          type: number
        status:
          type: string
          enum: [submitted, under_review, won, lost]
      required: [jobId, title, platform, bidAmount, status]
...
    AadhaarAuthRequest:
      type: object
      properties:
        aadhaarNumber:
          type: string
        otp:
          type: string
        purpose:
          type: string
      required: [aadhaarNumber, otp, purpose]
    AadhaarAuthResult:
      type: object
      properties:
        authenticated:
          type: boolean
        citizenId:
          type: string
        tokenExpiry:
          type: string
        permissionsGranted:
          type: array
          items:
            type: string
      required: [authenticated, permissionsGranted]
    UpiPaymentRequest:
      type: object
      properties:
        fromUpiId:
          type: string
        toUpiId:
          type: string
        amount:
          type: number
        purpose:
          type: string
        taskId:
          type: string
      required: [fromUpiId, toUpiId, amount, purpose]
    UpiPaymentResult:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [success, pending, failed]
        amount:
          type: number
        timestamp:
          type: string
        utrNumber:
          type: string
      required: [transactionId, status, amount, timestamp]
    DigiLockerDocs:
      type: object
      properties:
        citizenId:
          type: string
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DigiLockerDocument"
      required: [citizenId, documents]
    DigiLockerDocument:
      type: object
      properties:
        docId:
          type: string
        docType:
          type: string
        issuedBy:
          type: string
        issueDate:
          type: string
        verificationStatus:
          type: string
          enum: [verified, pending, expired]
      required: [docId, docType, issuedBy, issueDate, verificationStatus]
...